refactor: move tinyurl to its own router
small little refactor getting to a cleaner `views/api.py`
