+++ lnbits/core/__init__.py
@@ -8,6 +8,7 @@ from .views.api import api_router
 from .views.generic import generic_router, update_user_extension
 from .views.node_api import node_router, public_node_router, super_node_router
 from .views.public_api import public_router
+from .views.tinyurl_api import tinyurl_router
 
 # backwards compatibility for extensions
 core_app = APIRouter(tags=["Core"])
@@ -22,3 +23,4 @@ def init_core_routers(app):
     app.include_router(super_node_router)
     app.include_router(public_node_router)
     app.include_router(admin_router)
+    app.include_router(tinyurl_router)
